+++ dev/src/lib/tests/fields/lexical-editor-with-multiple-blocks.test.ts
@@ -1,6 +1,7 @@
 import payload from 'payload';
 import { initPayloadTest } from '../helpers/config';
 import {
+  getArticleDirectory,
   getFilesByDocumentID,
   isDefined,
   payloadCrowdinSyncTranslationsApi,
@@ -632,6 +633,45 @@ describe('Lexical editor with multiple blocks', () => {
     `);
   });
 
+  it('associates a parent Crowdin article directory with a lexical blocks Crowdin article directory', async () => {
+    nock('https://api.crowdin.com')
+      .post(`/api/v2/projects/${pluginOptions.projectId}/directories`)
+      .twice()
+      .reply(200, mockClient.createDirectory({}))
+      .post(`/api/v2/storages`)
+      .times(4)
+      .reply(200, mockClient.addStorage())
+      .post(`/api/v2/projects/${pluginOptions.projectId}/files`)
+      .times(4)
+      .reply(200, mockClient.createFile({}));
+
+    const create = await payload.create({
+      collection: 'policies',
+      data: {
+        title: 'Test policy',
+        content: fixture,
+      },
+    });
+    // update now that a Crowdin article directory is available
+    const policy: Policy = (await payload.update({
+      id: create.id,
+      collection: 'policies',
+      data: {
+        title: 'Test policy',
+      },
+    })) as any;
+    const lexicalBlocksArticleDirectory: CrowdinArticleDirectory =
+      (await getArticleDirectory(
+        `${pluginOptions.lexicalBlockFolderPrefix}content`,
+        payload,
+        false,
+        policy.crowdinArticleDirectory
+      )) as any;
+    expect(lexicalBlocksArticleDirectory).toBeDefined();
+    /** Important: ensure an article directory was not queried/returned without a parent */
+    expect(lexicalBlocksArticleDirectory?.parent).toBeDefined();
+  });
+
   it('creates HTML files for Crowdin as expected', async () => {
     nock('https://api.crowdin.com')
       .post(`/api/v2/projects/${pluginOptions.projectId}/directories`)
@@ -666,7 +706,11 @@ describe('Lexical editor with multiple blocks', () => {
     expect(contentHtmlFile?.fileData?.html).toMatchInlineSnapshot(
       `"<p>Sample content for a Lexical rich text field with multiple blocks.</p><span data-block-id=65d67d2591c92e447e7472f7></span><p>A bulleted list in-between some blocks consisting of:</p><ul class="bullet"><li value=1>one bullet list item; and</li><li value=2>another!</li></ul><span data-block-id=65d67d8191c92e447e7472f8></span><span data-block-id=65d67e2291c92e447e7472f9></span><ul class="bullet"><li value=1></li></ul>"`
     );
-    const contentBlocksCrowdinFiles = await getFilesByDocumentID(`${pluginOptions.lexicalBlockFolderPrefix}content`, payload, policy['crowdinArticleDirectory'] as CrowdinArticleDirectory);
+    const contentBlocksCrowdinFiles = await getFilesByDocumentID(
+      `${pluginOptions.lexicalBlockFolderPrefix}content`,
+      payload,
+      policy['crowdinArticleDirectory'] as CrowdinArticleDirectory
+    );
     //const contentBlocksHtmlFile = crowdinFiles.find(
     const contentBlocksHtmlFile = contentBlocksCrowdinFiles.find(
       (file) =>
@@ -690,7 +734,7 @@ describe('Lexical editor with multiple blocks', () => {
       .times(7)
       .reply(200, mockClient.createFile({}));
 
-    const policy: Policy = await payload.create({
+    const policy: Policy = (await payload.create({
       collection: 'policies',
       data: {
         group: {
@@ -706,7 +750,7 @@ describe('Lexical editor with multiple blocks', () => {
           ],
         },
       },
-    }) as any;
+    })) as any;
 
     const arrayField = isDefined(policy['group']?.['array'])
       ? policy['group']?.['array']
@@ -730,9 +774,16 @@ describe('Lexical editor with multiple blocks', () => {
       crowdinFiles.find((file) => file.name === 'fields.json')
     ).toBeDefined();
 
-
-    const fileOneCrowdinFiles = await getFilesByDocumentID(`${pluginOptions.lexicalBlockFolderPrefix}group.array.${ids[0]}.content`, payload, policy.crowdinArticleDirectory as CrowdinArticleDirectory);
-    const fileTwoCrowdinFiles = await getFilesByDocumentID(`${pluginOptions.lexicalBlockFolderPrefix}group.array.${ids[1]}.content`, payload, policy.crowdinArticleDirectory as CrowdinArticleDirectory);
+    const fileOneCrowdinFiles = await getFilesByDocumentID(
+      `${pluginOptions.lexicalBlockFolderPrefix}group.array.${ids[0]}.content`,
+      payload,
+      policy.crowdinArticleDirectory as CrowdinArticleDirectory
+    );
+    const fileTwoCrowdinFiles = await getFilesByDocumentID(
+      `${pluginOptions.lexicalBlockFolderPrefix}group.array.${ids[1]}.content`,
+      payload,
+      policy.crowdinArticleDirectory as CrowdinArticleDirectory
+    );
     expect(fileOneCrowdinFiles.length).toEqual(2);
     expect(fileTwoCrowdinFiles.length).toEqual(2);
 
@@ -971,7 +1022,11 @@ describe('Lexical editor with multiple blocks', () => {
       },
     });
     const crowdinFiles = await getFilesByDocumentID(`${policy.id}`, payload);
-    const contentCrowdinFiles = await getFilesByDocumentID(`${pluginOptions.lexicalBlockFolderPrefix}content`, payload, policy['crowdinArticleDirectory'] as CrowdinArticleDirectory);
+    const contentCrowdinFiles = await getFilesByDocumentID(
+      `${pluginOptions.lexicalBlockFolderPrefix}content`,
+      payload,
+      policy['crowdinArticleDirectory'] as CrowdinArticleDirectory
+    );
 
     // check file ids are always mapped in the same way
     const fileIds = crowdinFiles.map((file) => ({
@@ -1033,7 +1088,7 @@ describe('Lexical editor with multiple blocks', () => {
                   "version": 1,
                 },
               ],
-              "direction": null,
+              "direction": "ltr",
               "format": "",
               "indent": 0,
               "type": "paragraph",
@@ -1051,7 +1106,7 @@ describe('Lexical editor with multiple blocks', () => {
                   "version": 1,
                 },
               ],
-              "direction": null,
+              "direction": "ltr",
               "format": "",
               "indent": 0,
               "type": "paragraph",
@@ -1060,23 +1115,86 @@ describe('Lexical editor with multiple blocks', () => {
             {
               "children": [
                 {
-                  "detail": 0,
-                  "format": 0,
-                  "mode": "normal",
-                  "style": "",
-                  "text": "un élément de liste à puces ; etun autre!",
-                  "type": "text",
+                  "children": [
+                    {
+                      "detail": 0,
+                      "format": 0,
+                      "mode": "normal",
+                      "style": "",
+                      "text": "un élément de liste à puces ; et",
+                      "type": "text",
+                      "version": 1,
+                    },
+                  ],
+                  "direction": "ltr",
+                  "format": "",
+                  "indent": 0,
+                  "type": "listitem",
+                  "value": 1,
+                  "version": 1,
+                },
+                {
+                  "children": [
+                    {
+                      "detail": 0,
+                      "format": 0,
+                      "mode": "normal",
+                      "style": "",
+                      "text": "un autre!",
+                      "type": "text",
+                      "version": 1,
+                    },
+                  ],
+                  "direction": "ltr",
+                  "format": "",
+                  "indent": 0,
+                  "type": "listitem",
+                  "value": 2,
                   "version": 1,
                 },
               ],
-              "direction": null,
+              "direction": "ltr",
               "format": "",
               "indent": 0,
-              "type": "paragraph",
+              "listType": "bullet",
+              "start": 1,
+              "tag": "ul",
+              "type": "list",
+              "version": 1,
+            },
+            {
+              "children": [
+                {
+                  "children": [
+                    {
+                      "detail": 0,
+                      "format": 0,
+                      "mode": "normal",
+                      "style": "",
+                      "text": "",
+                      "type": "text",
+                      "version": 1,
+                    },
+                  ],
+                  "direction": "ltr",
+                  "format": "",
+                  "indent": 0,
+                  "type": "listitem",
+                  "value": 1,
+                  "version": 1,
+                },
+              ],
+              "direction": "ltr",
+              "format": "",
+              "indent": 0,
+              "listType": "bullet",
+              "start": 1,
+              "tag": "ul",
+              "type": "list",
               "version": 1,
             },
           ],
-          "direction": null,
+          "direction": "ltr",
           "format": "",
           "indent": 0,
           "type": "root",