+++ plugin/src/lib/utilities/richTextConversion.ts
@@ -17,10 +17,11 @@ import {
   consolidateHTMLConverters,
 } from '@payloadcms/richtext-lexical'
 
-import { createHeadlessEditor } from '@lexical/headless';
-import {$generateNodesFromDOM} from '@lexical/html';
-import { $getRoot, $getSelection } from 'lexical'
-import { Window } from 'happy-dom'
+import {
+  cloneDeep,
+  convertSlateToLexical,
+  defaultSlateConverters,
+} from '@payloadcms/richtext-lexical'
 
 const BlockHTMLConverter: HTMLConverter<any> = {
   converter: async ({ node }) => {
@@ -39,37 +40,16 @@ export const convertLexicalToHtml = async (editorData: SerializedEditorState, ed
   })
 }
 
-export const convertHtmlToLexical = async (htmlString: string, editorConfig: SanitizedEditorConfig) => {
-  const headlessEditor = createHeadlessEditor({ ...editorConfig, nodes: [] });
-  headlessEditor.update(() => {
-    
-    let document
-    if (process.env['NODE_ENV'] === "test") {
-      const contextWindow = new Window()
-      document = contextWindow.document;
-      document.body.innerHTML = htmlString;
-    } else {
-      const parser = new DOMParser();
-      document = parser.parseFromString(htmlString, "text/html");
-    }
-
-    // Once you have the DOM instance it's easy to generate LexicalNodes.
-    // type any required to pass Document type
-    // @see https://github.com/capricorn86/happy-dom/issues/1227
-    /* eslint-disable  @typescript-eslint/no-explicit-any */
-    const nodes = $generateNodesFromDOM(headlessEditor, document as any)
-  
-    // Select the root
-    $getRoot().select()
-  
-    // Insert them at a selection.
-    const selection = $getSelection()
-    selection?.insertNodes(nodes)
-  }, { discrete: true })
-
-  const editorJSON = headlessEditor.getEditorState().toJSON()
-
-  return editorJSON
+export const convertHtmlToLexical = (htmlString: string, editorConfig: SanitizedEditorConfig) => {
+  // use editorConfig to determine custom convertors
+  const converters = cloneDeep([
+    ...defaultSlateConverters,
+    // AnotherCustomConverter
+  ])
+  return convertSlateToLexical({
+    converters: converters,
+    slateData: convertHtmlToSlate(htmlString),
+  })
 }
 
 export const convertSlateToHtml = (slate: Descendant[], customConfig?: SlateToHtmlConfig): string => {