+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [0.7.8](https://github.com/thompsonsj/payload-crowdin-sync/compare/source-v0.7.7...source-v0.7.8) (2024-08-16)
+
+
+### Bug Fixes
+
+* **pluginfields:** additional flag for translation sync checkboxes ([#188](https://github.com/thompsonsj/payload-crowdin-sync/issues/188)) ([8480024](https://github.com/thompsonsj/payload-crowdin-sync/commit/84800242a4e9eea3962ed1d1e3999c6ba5efd75b))
+
+## [0.7.7](https://github.com/thompsonsj/payload-crowdin-sync/compare/source-v0.7.6...source-v0.7.7) (2024-08-13)
+
+
+### Bug Fixes
+
+* **richtextconversion:** use slate convertor to generate lexical ([#184](https://github.com/thompsonsj/payload-crowdin-sync/issues/184)) ([16fdc2a](https://github.com/thompsonsj/payload-crowdin-sync/commit/16fdc2a13d6926730f5326086851604596963d13))
+
+## [0.7.6](https://github.com/thompsonsj/payload-crowdin-sync/compare/source-v0.7.5...source-v0.7.6) (2024-07-26)
+
+
+### Bug Fixes
+
+* **utilities:** relocalize fn for tabs and collapsibles ([#182](https://github.com/thompsonsj/payload-crowdin-sync/issues/182)) ([de62e80](https://github.com/thompsonsj/payload-crowdin-sync/commit/de62e8029486fab9ff83a61928c60da278ed9da2))
+
+## [0.7.5](https://github.com/thompsonsj/payload-crowdin-sync/compare/source-v0.7.4...source-v0.7.5) (2024-07-26)
+
+
+### Bug Fixes
+
+* **bydocument:** lexical blocks article directory has parent ([#180](https://github.com/thompsonsj/payload-crowdin-sync/issues/180)) ([35d7fbd](https://github.com/thompsonsj/payload-crowdin-sync/commit/35d7fbd18b2fe7ace493be0e3b6016d36a6e941f))
+
 ## [0.7.4](https://github.com/thompsonsj/payload-crowdin-sync/compare/source-v0.7.3...source-v0.7.4) (2024-07-26)